Move directory parsing in preparation for multiple OS support.
