+++ kmeldb/mac_dir_parser.py
@@ -1,0 +1,2 @@
+def get_directory_entries(self, root, rootfd, files):
+    pass
